+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/combobox/DefaultComboBoxBuilder.java
@@ -47,7 +47,7 @@ public class DefaultComboBoxBuilder<T, C extends JComboBox<T>, B extends ComboBo
 
 	private boolean editable = false;
 	private Completion.Mode completionMode = Completion.COMPLETION_MODE.get();
-	private Completion.Normalize normalize = Completion.NORMALIZE.get();
+	private boolean normalize = Completion.NORMALIZE.get().equals(Completion.Normalize.YES);
 	private ListCellRenderer<T> renderer;
 	private ComboBoxEditor editor;
 	private boolean mouseWheelScrolling = true;
@@ -82,8 +82,8 @@ public class DefaultComboBoxBuilder<T, C extends JComboBox<T>, B extends ComboBo
 	}
 
 	@Override
-	public final B normalize(Completion.Normalize normalize) {
-		this.normalize = requireNonNull(normalize);
+	public final B normalize(boolean normalize) {
+		this.normalize = normalize;
 		return self();
 	}
 
@@ -148,7 +148,7 @@ public class DefaultComboBoxBuilder<T, C extends JComboBox<T>, B extends ComboBo
 			comboBox.setEditable(true);
 		}
 		if (!editable && editor == null) {
-			Completion.enable(comboBox, completionMode, normalize);
+			Completion.enable(comboBox, completionMode, normalize ? Completion.Normalize.YES : Completion.Normalize.NO);
 		}
 		if (comboBoxModel instanceof FilterComboBoxModel && comboBox.isEditable() && moveCaretToFrontOnSelection) {
 			((FilterComboBoxModel<T>) comboBoxModel).selection().item().addConsumer(new MoveCaretToStart<>(comboBox));